+++ CHANGELOG.md
@@ -78,6 +78,7 @@ This release contains several breaking changes to 1.6 due to heavy internal refa
 - Math doesn't render in comments anymore.
 - Opening files with Zettlr when the app is not running will now correctly open them.
 - Zooming on Windows and Linux can now be facilitated by scrolling while holding down the control-key.
+- Use `Cmd/Ctrl+Shift+L` to copy the active file's ID to the clipboard.
 
 ## Under the Hood